Initial commit
Create a new command-line tool for working with the Github CLI.

Initially inspired by hub.

I’d like to be able to query the Github API, without paging through
stuff. It would be lovely if they had a API browser, like the Guardian
Content API.

Until I create that for them, this is scratching my initial itch of
auditing everyone in an organisation, so that we can confirm that they
should still be members of the organisation.